Experiments with webgpu

+++ src/main.rs
@@ -1,5 +1,7 @@
+use dioxus::logger::tracing::{info, Level};
 use dioxus::prelude::*;
-use dioxus_logger::tracing::{info, Level};
+
+use crate::sections::ActiveSection;
 
 mod components;
 mod sections;
@@ -7,19 +9,74 @@ mod utils;
 
 #[cfg(debug_assertions)]
 fn config_logger() {
-    dioxus_logger::init(Level::DEBUG).expect("failed to init logger");
+    dioxus::logger::init(Level::DEBUG).expect("failed to init logger");
     info!("starting app");
 }
 
 #[cfg(not(debug_assertions))]
 fn config_logger() {
-    dioxus_logger::init(Level::WARN).expect("failed to init logger");
+    dioxus::logger::init(Level::WARN).expect("failed to init logger");
     info!("starting app");
 }
 
-#[cfg(not(target_family = "wasm"))]
+#[cfg(feature = "generate_htmls")]
+fn generate_blog_files(docs_dir: &std::path::Path, index_file: &std::path::PathBuf) {
+    let blog_dir = docs_dir.join("blog");
+    for entry in std::fs::read_dir("assets/blog").unwrap() {
+        let origin_path_year = entry.unwrap().path();
+        if !origin_path_year.is_dir() {
+            continue;
+        }
+
+        let year_dir = blog_dir.join(origin_path_year.file_name().unwrap());
+        if !year_dir.exists() {
+            std::fs::create_dir_all(year_dir.clone())
+                .expect("failed to create year_dir dir for a post");
+        } else {
+            assert!(year_dir.is_dir());
+        }
+        for entry in origin_path_year.read_dir().unwrap() {
+            let origin_path_month = entry.unwrap().path();
+            if !origin_path_month.is_dir() {
+                continue;
+            }
+
+            let month_dir = year_dir.join(origin_path_month.file_name().unwrap());
+            if !month_dir.exists() {
+                std::fs::create_dir_all(month_dir.clone())
+                    .expect("failed to create month_dir dir for a post");
+            } else {
+                assert!(month_dir.is_dir());
+            }
+            for entry in origin_path_month.read_dir().unwrap() {
+                let origin_path_day = entry.unwrap().path();
+                if !origin_path_day.is_dir() {
+                    continue;
+                }
+
+                let day_dir = month_dir.join(origin_path_day.file_name().unwrap());
+                if !day_dir.exists() {
+                    std::fs::create_dir_all(day_dir.clone())
+                        .expect("failed to create day_dir dir for a post");
+                } else {
+                    assert!(day_dir.is_dir());
+                }
+
+                let file = day_dir.join("index.html");
+                let file_str = file.to_str().unwrap();
+                info!("file: {file_str}");
+
+                std::fs::copy(index_file, file).expect("failed to copy index.html to new route");
+            }
+        }
+    }
+}
+
+#[cfg(feature = "generate_htmls")]
 fn generate_all_route_files() {
-    let docs_dir = std::env::current_dir().unwrap().join("docs");
+    let docs_dir = std::env::current_dir()
+        .unwrap()
+        .join(std::env::args().collect::<Vec<_>>().get(1).unwrap());
     assert!(
         docs_dir.exists(),
         "docs directory not found. Please run `./build.sh`"
@@ -33,7 +90,7 @@ fn generate_all_route_files() {
 
     for route in sections::ActiveSection::all_routes() {
         let route = route.join("/");
-        let file = docs_dir.join(route);
+        let file = docs_dir.join(&route);
         let parent_dir = file.parent().unwrap();
         let file_str = file.to_str().unwrap();
         info!("file: {file_str}");
@@ -43,21 +100,72 @@ fn generate_all_route_files() {
             assert!(parent_dir.is_dir());
         }
 
-        std::fs::copy(&index_file, file).expect("failed to copy index.html to new route");
+        let static_version = std::env::current_dir()
+            .unwrap()
+            .join("static")
+            .join(route)
+            .join("index.html");
+
+        if static_version.is_file() {
+            info!(
+                "found static version at: {}",
+                static_version.to_str().unwrap()
+            );
+            std::fs::copy(static_version, file)
+                .expect("failed to copy static version of page to new route");
+        } else {
+            std::fs::copy(&index_file, file).expect("failed to copy index.html to new route");
+        }
+    }
+
+    // Special case for posts
+    generate_blog_files(&docs_dir, &index_file);
+
+    // Special case for the root
+    let static_version = std::env::current_dir()
+        .unwrap()
+        .join("static")
+        .join("index.html");
+
+    if static_version.is_file() {
+        info!(
+            "found static version at: {}",
+            static_version.to_str().unwrap()
+        );
+        std::fs::copy(static_version, index_file)
+            .expect("failed to copy static version of page to new route");
     }
 }
 
-#[cfg(not(target_family = "wasm"))]
+#[cfg(feature = "generate_htmls")]
 fn main() {
     config_logger();
     generate_all_route_files();
 }
 
-#[cfg(target_family = "wasm")]
+#[cfg(any(feature = "server", feature = "web"))]
 fn main() {
     // Init logger
     config_logger();
+    dioxus::LaunchBuilder::new()
+        .with_cfg(server_only! {
+            ServeConfig::builder()
+                // turn on incremental site generation with the .incremental() method
+                .incremental(IncrementalRendererConfig::new())
+                .build()
+                .unwrap()
+        })
+        .launch(App);
     launch(App);
+}
+
+#[server(endpoint = "static_routes")]
+async fn static_routes() -> Result<Vec<String>, ServerFnError> {
+    Ok(ActiveSection::all_static_routes()
+        .into_iter()
+        .map(|r| format!("/{}", r.join("/")))
+        .chain(vec!["/".to_string()].into_iter())
+        .collect())
 }
 
 #[component]
@@ -69,10 +177,15 @@ fn MainSectionDisplayed(route: Vec<String>, current_section: sections::ActiveSec
         sections::ActiveSection::PasswordGenerator => rsx! {
             main { sections::password_generator::PasswordGenerator {} }
         },
+        sections::ActiveSection::Blog(blog_date) => rsx! {
+            main {
+                sections::blog::Blog { blog_date }
+            }
+        },
     }
 }
 
-#[derive(Routable, Clone)]
+#[derive(Debug, Routable, Clone)]
 #[rustfmt::skip]
 pub enum Route {
     //#[route("/:..route?:..query")]
@@ -86,6 +199,8 @@ pub enum Route {
 pub fn App() -> Element {
     info!("Starting app");
     rsx! {
+        document::Stylesheet { href: asset!("assets/css/tailwind.css") }
+        document::Stylesheet { href: asset!("assets/css/hamburgers-min.css") }
         Router::<Route> {}
     }
 }
@@ -106,6 +221,11 @@ fn NavBar(route: Vec<String>) -> Element {
         components::title_bar::TitleEntry::new(
             "Password Generator",
             sections::ActiveSection::PasswordGenerator,
+            current_section,
+        ),
+        components::title_bar::TitleEntry::new(
+            "Blog",
+            sections::ActiveSection::Blog(None),
             current_section,
         ),
     ];
